<map> changing param name to be compatible with iOS
git-svn-id: file:///var/pocketcampus/trunk@1428 fd477b11-563a-408f-9d6e-eeb9ca009d0c

+++ plugin/map/plugin.map.server/src/org/pocketcampus/plugin/map/server/MapServiceImpl.java
@@ -42,9 +42,8 @@ public class MapServiceImpl implements MapService.Iface {
 	}
 
 	@Override
-	public List<MapItem> getLayerItems(int layerId) throws TException {
-		mItemsList = mMapDb.getMapElements(layerId);
-		
+	public List<MapItem> getLayerItems(long layerId) throws TException {
+		mItemsList = mMapDb.getMapElements((int) layerId);
 		return mItemsList;
 	}
